feat: moved logic from RosaConnector to RosaIntegration and make function calls from RiScController
