+++ .env.example
@@ -5,6 +5,7 @@ RISC_FOLDER_PATH=".security/risc"
 FILENAME_PREFIX="risc"
 FILENAME_POSTFIX="risc"
 CRYPTO_SERVICE_URL="http://localhost:8084"
+ROSA_URL="http://localhost:8888"
 INIT_RISC_URL="http://localhost:8085"
 
 GITHUB_APP_ID=... # App-ID of the GitHub app used by users of RiSc-plugin without GitHub account.
