+++ src/main/kotlin/no/risc/config/RosaConfig.kt
@@ -1,0 +1,10 @@
+package no.risc.config
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.stereotype.Component
+
+@Component
+@ConfigurationProperties(prefix = "rosa")
+class RosaConfig {
+    lateinit var baseUrl: String
+}
