+++ src/main/resources/application.properties
@@ -14,6 +14,7 @@ management.endpoint.health.access=read_only
 management.endpoint.prometheus.access=read_only
 management.endpoints.web.exposure.include=prometheus,health
 cryptoService.baseUrl=${CRYPTO_SERVICE_URL}
+rosa.baseUrl=${ROSA_URL}
 initRisc.baseUrl=${INIT_RISC_URL}
 sops.backendPublicKey=${BACKEND_PUBLIC_KEY}
 github.app.private-key=${GITHUB_APP_PRIVATE_KEY}
